Refactor: delete item functional

+++ src/pages/doador/animais-cadastrados.tsx
@@ -1,17 +1,20 @@
 import { Footer } from "@/components/Footer";
+import { myFont } from "@/components/pages/Home/Hero";
+import CardAnimal from "@/components/pages/doador/CardAnimal";
 import Sidebar from "@/components/pages/doador/Sidebar";
 import { authValidate } from "@/utils/authUtils";
+import { toastActive } from "@/utils/toastComponent";
+import { format, formatISO } from "date-fns";
 import { GetServerSideProps } from "next";
 import { Inter } from "next/font/google";
 import Head from "next/head";
-import { prisma } from "../../../lib/prisma";
-import { format, formatISO } from "date-fns";
-import CardAnimal from "@/components/pages/doador/CardAnimal";
-import { api } from "../../../lib/axios";
-import { useState } from "react";
+import Image from "next/image";
+import { useEffect, useState } from "react";
 import { ToastContainer } from "react-toastify";
-import { toastActive } from "@/utils/toastComponent";
+import { api } from "../../../lib/axios";
+import { prisma } from "../../../lib/prisma";
 const inter = Inter({ subsets: ["latin"] });
+import ConfusedCat from "../../assets/images/confusedCat.png";
 
 interface PropsFormDoacao {
   id: string;
@@ -37,27 +40,41 @@ interface AnimalCadastrado {
 }
 
 function AnimaisCadastrados({ dataAnimal }: AnimalCadastrado) {
-  const [listAnimals, setListAnimals] = useState(dataAnimal);
   const [idsDeleted, setIdDeleted] = useState<string[]>([]);
+  const [isSaving, setIsSaving] = useState(false);
 
   function handleDeletePet(id: string) {
-    const listAnimalsUpdated = listAnimals.filter((animal) => animal.id !== id);
-    setListAnimals(listAnimalsUpdated);
-    setIdDeleted((prevState) => [...prevState, id]);
+    setIdDeleted((prevState) => {
+      if (prevState.includes(id)) {
+        return prevState.filter((item) => item !== id);
+      } else {
+        return [...prevState, id];
+      }
+    });
   }
 
+  useEffect(() => {
+    console.log(idsDeleted);
+  }, [idsDeleted]);
   async function handleModifyData() {
-    try {
-      const res = await api.delete("/upload", {
-        params: {
-          ids: idsDeleted
+    if (dataAnimal.length !== 0 && idsDeleted.length !== 0) {
+      try {
+        setIsSaving(true);
+        const res = await api.delete("/upload", {
+          params: {
+            ids: idsDeleted,
+          },
+        });
+        if (res.status === 200) {
+          await toastActive({ error: false, message: "Animal deletado" });
+          setIsSaving(false);
+          setIdDeleted([]);
         }
-      });
-      if(res.status === 200){
-        await toastActive({error: false, message: "Animal deletado"});
+      } catch (error) {
+        await toastActive({ error: true, message: "Erro ao deletar animal" });
       }
-    } catch (error) {
-      await toastActive({error: true, message: "Erro ao deletar animal"});
+    } else {
+      await toastActive({ error: true, message: "Sem registros para deletar" });
     }
   }
   return (
@@ -76,23 +93,42 @@ function AnimaisCadastrados({ dataAnimal }: AnimalCadastrado) {
         <Sidebar activeMenu="animais-cadastrados" />
         <div className="w-full">
           <div className="flex gap-3 items-start flex-col">
-            {listAnimals.map(({ id, name, city, details, photos }) => (
-              <CardAnimal
-                setDeletePet={handleDeletePet}
-                city={city}
-                details={details}
-                photos={photos}
-                name={name}
-                key={id}
-                id={id}
-              />
-            ))}
+            {dataAnimal.length >= 1 ? (
+              dataAnimal.map(({ id, name, city, details, photos }) => (
+                <CardAnimal
+                  setDeletePet={handleDeletePet}
+                  city={city}
+                  details={details}
+                  photos={photos}
+                  name={name}
+                  key={id}
+                  id={id}
+                />
+              ))
+            ) : (
+              <div className="flex flex-col items-center justify-center w-full">
+                <p className={`${myFont.className} text-dark-blue text-2xl`}>
+                  Nenhum animal Cadastrado.
+                </p>
+                <div className="relative w-[300px] h-[300px]">
+                  <Image
+                    alt=""
+                    src={ConfusedCat}
+                    width={300}
+                    height={300}
+                    quality={100}
+                    priority
+                  />
+                </div>
+              </div>
+            )}
           </div>
           <button
             onClick={handleModifyData}
-            className="bg-black text-xl text-white p-2 rounded mt-4"
+            disabled={isSaving}
+            className="bg-purple-500 text-white px-4 py-2 rounded hover:bg-purple-700 transition-all block ml-auto mt-8 disabled:cursor-not-allowed disabled:bg-purple-950"
           >
-            Salvar
+            {isSaving ? "Salvando" : "Salvar"}
           </button>
         </div>
       </main>
